Move ConditionalWrap
