+++ packages/ui/src/ValueViewComponent/index.tsx
@@ -1,5 +1,5 @@
 import { ValueView } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/asset/v1/asset_pb';
-import { ConditionalWrap } from '../ConditionalWrap';
+import { ConditionalWrap } from '../utils/ConditionalWrap';
 import { Pill } from '../Pill';
 import { Text } from '../Text';
 import styled from 'styled-components';
